Improve code quality with pylint
