+++ requirements.txt
@@ -4,7 +4,7 @@ Twisted==15.0.0
 argparse==1.2.1
 cffi==0.9.1
 characteristic==14.3.0
-cryptography==0.8
+cryptography==1.3.1
 cssselect==0.9.1
 enum34==1.0.4
 html2text==2015.2.18
